pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "inboxzero-tool"
version = "1.0.0"
description = "A powerful desktop tool to manage Gmail labels, filters, and achieve Inbox Zero."
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
  {name = "Tu Nombre", email = "tu@email.com"}
]
dependencies = [
    "google-api-python-client>=2.129.0",
    "google-auth>=2.33.0",
    "google-auth-oauthlib>=1.2.1",
    "cryptography>=41.0.0"
]
keywords = ["gmail", "email", "management", "cleanup", "inbox-zero"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
]

[project.urls]
"Homepage" = "https://github.com/tu-usuario/inboxzero-tool"
"Bug Tracker" = "https://github.com/tu-usuario/inboxzero-tool/issues"

[project.scripts]
inboxzero = "gmail_manager.main:main"

[tool.setuptools.packages.find]
where = ["src"]
